Callback classes should disappear so make them inner classes of CounterHandler
CounterHandler is the last handler using them
